+++ src/plugins/Dropbox.js
@@ -3,7 +3,7 @@ import Plugin from './Plugin'
 export default class Dropbox extends Plugin {
   constructor (core, opts) {
     super(core, opts)
-    this.type = 'selecter'
+    this.type = 'acquire'
     this.authenticate = this.authenticate.bind(this)
     this.connect = this.connect.bind(this)
     this.render = this.render.bind(this)