+++ test/mocks/plugin-acquire1.js
@@ -1,9 +1,9 @@
 const Plugin = require('../../src/plugins/Plugin.js')
 
-export default class TestSelector2 extends Plugin {
+export default class TestSelector1 extends Plugin {
   constructor (core, opts) {
     super(core, opts)
-    this.type = 'selecter'
+    this.type = 'acquire'
     this.name = this.constructor.name
   }